Fix glfw source path
